requirements.txt
pandas>=1.5.0
plotly>=5.13.0
google-cloud-bigquery>=3.0.0 